Squash bug/fix-spiderfied-state-on-update-layers (#2)
Added the deletion of the current `_spiderfied` MarkerCluster on `_noanimationUnspiderfy()`.

If this is not done and `spiderfy` or `noanimationSpiderfy` are called afterwards, these methods will try to `unspiderfy` the current `_spiderfied` MarkerCluster. That results in javascript errors and the functionality to spiderfy MarkerCluster's is broken until the next page reload.